pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "khoregos"
version = "0.1.0"
description = "Enterprise governance layer for Claude Code Agent Teams"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Davy Braun" }
]
keywords = ["claude", "agents", "governance", "audit", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "watchdog>=4.0",
    "aiosqlite>=0.19",
    "mcp>=1.0",
    "python-ulid>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "mypy>=1.8",
    "ruff>=0.2",
]

[project.scripts]
k6s = "k6s.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/k6s"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]